fix(ui,api): address CodeRabbit review feedback on redesign stage 2
- Require painLevel as an integer, matching the Prisma field.
- Surface load errors instead of showing empty/wrong states (timeline,
  follow-ups, treatment outcomes, injury detail).
- Fix logout: only redirect after it actually succeeds.
- Fix two stale-response races with a request-generation counter
  (injuries list, treatment outcomes).
- Scope treatment comparison grouping to an injury, not just name.
- Validate ?injuryId= before using it as a form default.
- Fix follow-up date math to use calendar days, not fixed milliseconds.

Not fixed: the injury start-date display bug traces to a write-side
issue outside this PR (create-injury-dialog.tsx). Fixed the display
only; the write-side bug needs its own issue.

Verified: backend 52/52, frontend lint clean, frontend build passes.

+++ backend/src/validators.js
@@ -102,7 +102,7 @@ export const treatmentOutcomeSchema = z.object({
 
   reliefDays: z.number().int().nonnegative().optional().nullable(),
 
-  painLevel: z.number().min(0).max(10).optional().nullable(),
+  painLevel: z.number().int().min(0).max(10).optional().nullable(),
 
   notes: z.string().min(1).optional().nullable(),
 });
